+++ bindings/core/src/method/account.rs
@@ -18,7 +18,7 @@ use iota_sdk::{
     },
     wallet::{
         account::{
-            CreateAliasParamsDto, FilterOptions, MintNativeTokenParamsDto, MintNftParamsDto, OutputParamsDto,
+            CreateAliasParams, FilterOptions, MintNativeTokenParamsDto, MintNftParamsDto, OutputParamsDto,
             OutputsToClaim, SyncOptions, TransactionOptionsDto,
         },
         SendAmountParams, SendNativeTokensParams, SendNftParams,
@@ -152,7 +152,7 @@ pub enum AccountMethod {
     /// Create an alias output.
     /// Expected response: [`PreparedTransaction`](crate::Response::PreparedTransaction)
     PrepareCreateAliasOutput {
-        params: Option<CreateAliasParamsDto>,
+        params: Option<CreateAliasParams>,
         options: Option<TransactionOptionsDto>,
     },
     /// Melt native tokens. This happens with the foundry output which minted them, by increasing it's