+++ bindings/nodejs/lib/secret_manager/index.ts
@@ -1,0 +1,6 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './secret-manager';
+export * from './secret-manager-method-handler';
+export * from '../types/secret_manager';