+++ bindings/nodejs/lib/types/block/input/input.ts
@@ -1,0 +1,67 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+import { HexEncodedString } from '../../utils';
+
+/**
+ * All of the input types.
+ */
+enum InputType {
+    UTXO = 0,
+    Treasury = 1,
+}
+
+abstract class Input {
+    private type: InputType;
+
+    constructor(type: InputType) {
+        this.type = type;
+    }
+
+    /**
+     * The type of input.
+     */
+    getType(): InputType {
+        return this.type;
+    }
+}
+
+/**
+ * Treasury Input.
+ */
+class TreasuryInput extends Input {
+    /**
+     * The milestone id of the input.
+     */
+    milestoneId: HexEncodedString;
+
+    constructor(milestoneId: HexEncodedString) {
+        super(InputType.Treasury);
+        this.milestoneId = milestoneId;
+    }
+}
+
+/**
+ * UTXO Transaction Input.
+ */
+class UTXOInput extends Input {
+    /**
+     * The transaction Id.
+     */
+    transactionId: HexEncodedString;
+    /**
+     * The input index.
+     */
+    transactionInputIndex: number;
+
+    constructor(
+        transactionId: HexEncodedString,
+        transactionInputIndex: number,
+    ) {
+        super(InputType.Treasury);
+        this.transactionId = transactionId;
+        this.transactionInputIndex = transactionInputIndex;
+    }
+}
+
+export { InputType, Input, TreasuryInput, UTXOInput };