+++ bindings/nodejs/lib/types/client/network.ts
@@ -1,7 +1,7 @@
 // Copyright 2021-2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import type { INodeInfoProtocol } from '@iota/types';
+import { INodeInfoProtocol } from '../models/info';
 
 /**
  * Network types.