+++ bindings/nodejs/lib/types/index.ts
@@ -3,6 +3,15 @@
 
 export * from './block';
 export * from './client';
-export * from './secretManager';
+export * from './models';
+export * from './secret_manager';
 export * from './utils';
 export * from './wallet';
+
+/**
+ * Response from the message interface
+ */
+export interface Response<T> {
+    type: string;
+    payload: T;
+}