+++ sdk/src/wallet/account/builder.rs
@@ -44,8 +44,8 @@ where
     }
 
     /// Set the alias
-    pub fn with_alias(mut self, alias: impl Into<Option<String>>) -> Self {
-        self.alias = alias.into();
+    pub fn with_alias(mut self, alias: impl Into<String>) -> Self {
+        self.alias = Some(alias.into());
         self
     }